add collision_handler.py and play_area.py
